grok/requirements.txt
xai-sdk[telemetry-http]
opentelemetry-sdk
opentelemetry-exporter-otlp
